+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: `Button`: Color icons will not be styled by Button SVG styles.
+
 ## [3.15.0] - 2021-05-28
 
 Features: